+++ packages/react-app/src/components/Header.jsx
@@ -3,21 +3,16 @@ import React from "react";
 
 // displays a page header
 
-export default function Header({link, title, subTitle}) {
+export default function Header({ link, title, subTitle }) {
   return (
     <a href={link} target="_blank" rel="noopener noreferrer">
-      <PageHeader
-        title={title}
-        subTitle={subTitle}
-        style={{ cursor: "pointer" }}
-      />
+      <PageHeader title={title} subTitle={subTitle} style={{ cursor: "pointer" }} />
     </a>
   );
 }
 
-
 Header.defaultProps = {
   link: "https://github.com/austintgriffith/scaffold-eth",
-  title: "🏗 scaffold-eth",
-  subTitle: "forkable Ethereum dev stack focused on fast product iteration",
-}
+  title: "Gitcoin Conviction Voting",
+  subTitle: "Signal your support to the community",
+};
