+++ packages/react-app/src/views/index.js
@@ -1,4 +1,4 @@
 export { default as Home } from "./Home";
-export { default as ExampleUI } from "./ExampleUI";
-export { default as Hints } from "./Hints";
-export { default as Subgraph } from "./Subgraph";
+export { default as Checkout } from "./Checkout";
+export { default as Confirm } from "./Confirm";
+export { default as Success } from "./Success";
